\name{denoiseR-package}
\alias{denoiseR-package}
\alias{denoiseR}
\docType{package}
\title{
Regularized Low Rank Matrix Estimation
}
\description{
The methods implemented allow to recover a low-rank structure from noisy data. In addition, they may be used to estimate the underlying rank and to impute missing values.
}
\details{
\tabular{ll}{
Package: \tab denoiseR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-07-09\cr
License: \tab  GPL (>=2)\cr
}
}
\author{
Julie Josse, Sylvain Sardy, Stefan Wager \cr
Maintainer: Julie Josse <julie.josserennes@gmail.com>\cr
}
\references{Julie Josse, Sylvain Sardy, Stefan Wager. denoiseR a package for low rank matrix estimation.
}
\keyword{ package }
\seealso{
URL: http://juliejosse.com/\cr
http://web.stanford.edu/~swager/research.html\cr
http://www.unige.ch/math/folks/sardy\cr
}

