% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize.R
\name{finalize}
\alias{finalize}
\title{Finalize data after analysis}
\usage{
finalize(data, cols = NULL, .by = NULL)
}
\arguments{
\item{data}{Dataframe that contains information on location (x- and 
y-coordinates) and the time at which the measurement was taken. By default, 
[kalman_filter()] will assume that this information is contained within the 
columns `"x"`, `"y"`, and `"time"` respectively.}

\item{cols}{Named vector or named list containing the relevant column names
in 'data' if they didn't contain the prespecified column names `"time"`, 
`"x"`, and `"y"`. The labels should conform to these prespecified 
column names and the values given to these locations should contain the 
corresponding column names in that dataset. Defaults to `NULL`, therefore 
assuming the structure explained in `data`.}

\item{.by}{String denoting whether the moving window should be taken with 
respect to a given grouping variable. Defaults to `NULL`.}
}
\value{
Adjusted `data.frame`
}
\description{
Function that is used internally to change the data back to their original 
format after conducting an analysis, it being either binning ([bin()]) or 
using the Kalman filter ([kalman_filter()]). The finalization consists of 
renaming the columns to the user-defined column names instead of the package-
required ones and deleting the grouping variable if it was not originally 
present in the data.
}
\examples{
# Generate data for illustration purposes
#
# Note that for this to work, I need to define the package-required column 
# names
data <- data.frame(
  x = rnorm(100),
  y = rnorm(100),
  time = rep(1:50, times = 2) / 10,
  id = rep(1:2, each = 50)
)

# Prepare the data for analysis
finalize(
  data,
  cols = c(
    "time" = "seconds",
    "x" = "X",
    "y" = "Y"
  ),
  .by = "tag"
)

}
