% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement_models.R
\name{independent}
\alias{independent}
\title{Add independent error to data}
\usage{
independent(
  data,
  mean = c(0, 0),
  covariance = matrix(c(0.031^2, 0.02, 0.02, 0.027^2), nrow = 2, ncol = 2)
)
}
\arguments{
\item{data}{A data.frame containing the data on which to base the parameters
of the constant velocity model. This function assumes that this data.frame
contains the columns \code{"time"}, \code{"x"}, and \code{"y"} containing 
the time at which the observed position (x, y) was measured respectively.}

\item{mean}{Numeric vector with 1 or 2 values containing the mean of the 
error process. If only 1 number is provided, then this value will be assumed
to represent the mean for both the x- and y-dimensions. Defaults to 
\code{c(0, 0)}, representing no bias in the system.}

\item{covariance}{Numeric matrix of dimensionality \eqn{2 \times 2}, 
having the variances of the errors for x and y on the diagonal and the 
covariance of the errors on the off-diagonal. Defaults to having the variances
of the x- and y-dimension being equal to \code{0.031^2} and \code{0.027^2}
respectively, and the covariance being equal to \code{0.02}. These values 
represent the means of the respective variances and covariances found 
across the four days of the calibration study that informed this project.}
}
\value{
A \code{data.frame} in which the \code{"x"} and \code{"y"} columns 
have additional noise in them.
}
\description{
Error is generated through a multivariate normal distribution with a
\code{mean} and \code{covariance} that are specified by the user. Primary 
assumption of this model is that the errors are independent over time, 
which stands in contrast to the \code{\link[denoiser]{temporal()}} model.
}
\examples{
# Generate data for illustration purposes. Movement in circular motion at a
# pace of 1.27m/s with some added noise of SD = 10cm.
angles <- seq(0, 4 * pi, length.out = 100)
coordinates <- 10 * cbind(cos(angles), sin(angles))

data <- data.frame(
  x = coordinates[, 1],
  y = coordinates[, 2],
  time = 1:100
)

# Add independent noise to these data with an error variance of 0.01 in each
# dimension
independent(
  data,
  mean = c(0, 0),
  covariance = diag(2) * 0.01
) |>
  head()

}
\seealso{
\code{\link[denoiser]{noiser()}}
\code{\link[denoiser]{temporal()}}
}
