% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_models.R
\docType{data}
\name{kalman_models}
\alias{kalman_models}
\title{List of Kalman filter models}
\format{
An object of class \code{list} of length 2.
}
\usage{
kalman_models
}
\description{
List of Kalman filter models
}
\keyword{datasets}
