% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_filter.R
\name{kf_predict}
\alias{kf_predict}
\title{Predict step in the Kalman filter}
\usage{
kf_predict(
  x0,
  P0,
  F,
  W,
  u = matrix(0, nrow = length(x0), ncol = 1),
  B = matrix(0, nrow = length(u), ncol = length(u))
)
}
\arguments{
\item{x0}{Vector of values of the movement equation at time t.}

\item{P0}{Cholesky decomposition of the covariance of x0 at time t.}

\item{F}{Transition matrix, relating values of X at time t to those at t.}

\item{W}{Cholesky decomposition of the process noise covariance matrix.}

\item{u}{Vector of values for the external variables at time t. By default an 
empty vector.}

\item{B}{Matrix that connects the external variables in u to the values in x.
By default an empty matrix.}
}
\value{
List of predicted values for x and P at time t + 1.
}
\description{
In this step, we use the movement equation to predict the new values of x 
based on the initial condition \code{x0}, namely through the equation
}
\details{
\begin{equation}
    x = F x0 + G u
\end{equation}
where F is the movement transition matrix, G is the error matrix, and u is 
the error at that time. For the constant velocity model, this error is 
equal to random changes in acceleration.

We also compute the certainty around this prediction based on the initial 
estimate of this covariance and updated with the error variance associated to 
u, so that:

\begin{equation}
    P = F P0 F^T + G U G^T
\end{equation}
where U is the covariance matrix of u. We additionally contract $G u$ and 
$G U G^T$ into the singular error $w$ and its covariance matrix $W$ for 
simplicity purposes.

Note that in this implementation, we always use Cholesky decomposition 
instead of actual covariances.
}
