% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_filter.R
\name{kf_update}
\alias{kf_update}
\title{Update step in the Kalman filter}
\usage{
kf_update(x, P, y, H, K)
}
\arguments{
\item{x}{Vector of predicted values of the movement equation at time t + 1.}

\item{P}{Cholesky decomposition of the estimation covariance matrix at time 
t + 1.}

\item{y}{Vector of innovations at time t + 1}

\item{H}{Measurement matrix connecting measurement to movement.}

\item{K}{Kalman gain}
}
\value{
List containing the smoothed value of x ("x") together with the  
Cholesky decomposition of its covariance ("P")
}
\description{
In the update step, we use the measurement and the predicted value of the 
variable x and use both to make a good guess of what the "real" value of 
x might be (as well as what the uncertainty around this guess is).
}
\details{
We do this using the following equation:

\begin{equation}
    x_t = x + K y
\end{equation}
where x is the predicted value for the variable (coming from the prediction 
step), K is the Kalman gain, and y is the innovation (both coming from the 
innovation step).

For the covariance, we use the following equation:

\begin{equation}
    P_t = (I - K H) P
\end{equation}
where I is the identity matrix and P is the prediction covariance coming from 
the prediction step.
}
