% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement_models.R
\docType{data}
\name{measurement_models}
\alias{measurement_models}
\title{List of measurement models}
\format{
An object of class \code{list} of length 2.
}
\usage{
measurement_models
}
\description{
List of measurement models
}
\keyword{datasets}
