% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement_models.R
\name{temporal}
\alias{temporal}
\title{Add temporal error to data}
\usage{
temporal(
  data,
  intercept = c(0, 0),
  transition = matrix(c(0.925, 0.085, 0.085, 0.87), nrow = 2, ncol = 2),
  covariance = matrix(c(0.031^2, 0.02, 0.02, 0.027^2), nrow = 2, ncol = 2),
  sampling_rate = 6.13
)
}
\arguments{
\item{data}{A data.frame containing the data on which to base the parameters
of the constant velocity model. This function assumes that this data.frame
contains the columns \code{"time"}, \code{"x"}, and \code{"y"} containing 
the time at which the observed position (x, y) was measured respectively.}

\item{intercept}{Numeric vector with 1 or 2 values containing the intercept 
of the vector autoregressive model.If only 1 number is provided, then this 
value will be assumed to represent the intercept for both the x- and y-
dimensions. Defaults to \code{c(0, 0)}, representing no bias in the system.}

\item{transition}{Numeric matrix of dimensionality \eqn{2 \times 2} having 
the autoregressive effects on the diagonal and crossregressive effects on the
off-diagonal. Defaults to the autoregressive effects being equal to 
\code{0.925} and \code{0.87} for the x- and y-dimension respectively, and 
the cross-regressive effects being equal to \code{0.085} in a symmetric way.
These values represent the means of these parameters as found across the 
four days of the calibration study that informed this project.}

\item{covariance}{Numeric matrix of dimensionality \eqn{2 \times 2}, 
having the variances of the errors for x and y on the diagonal and the 
covariance of the errors on the off-diagonal. Defaults to having the variances
of the x- and y-dimension being equal to \code{0.031^2} and \code{0.027^2}
respectively, and the covariance being equal to \code{0.02}. These values 
represent the means of the respective variances and covariances found 
across the four days of the calibration study that informed this project.}

\item{sampling_rate}{Numeric denoting the sampling rate in Hz. Is used to 
scale the transition matrix according to the actual sampling rate. Defaults to 
\code{6.13}, which is the mean sampling rate for the data obtained across the
four days of the calibration study that informed this project. If you change
the \code{transition} argument, it is recommended to ignore this argument.}
}
\value{
A \code{data.frame} in which the \code{"x"} and \code{"y"} columns 
have additional noise in them.
}
\description{
Error is generated through a vector autoregressive model with defining
parameters \code{intercept}, \code{transition}, and \code{covariance} being
specified by the user.
}
\examples{
# Generate data for illustration purposes. Movement in circular motion at a
# pace of 1.27m/s with some added noise of SD = 10cm.
angles <- seq(0, 4 * pi, length.out = 100)
coordinates <- 10 * cbind(cos(angles), sin(angles))

data <- data.frame(
  x = coordinates[, 1],
  y = coordinates[, 2],
  time = 1:100
)

# Add independent noise to these data with an error variance of 0.01 in each
# dimension and a transition matrix containing autoregressive effects of 0.15
temporal(
  data,
  intercept = c(0, 0),
  transition = diag(2) * 0.15,
  covariance = diag(2) * 0.01
) |>
  head()

}
\seealso{
\code{\link[denoiser]{noiser()}}
\code{\link[denoiser]{independent()}}
}
