% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_models.R
\name{constant_velocity}
\alias{constant_velocity}
\title{Constant velocity model}
\usage{
constant_velocity(data, error = 0.031^2)
}
\arguments{
\item{data}{A data.frame containing the data on which to base the parameters
of the constant velocity model. This function assumes that this data.frame
contains the columns \code{"time"}, \code{"x"}, and \code{"y"} containing 
the time at which the observed position (x, y) was measured respectively.}

\item{error}{Numeric or numerical vector containing the assumed
value of the measurement error variance in the x- and y-direction. Should 
consist of either 1 or 2 values. If only 1 value is provided, the measurement
error variance will be assumed to be the same for both dimensions. Defaults 
to \code{0.031^2}, a value that we have obtained experimentally.}
}
\value{
Named list containing all parameters relevant for the Kalman filter.
}
\description{
This model assumes that movement occurs at a constant velocity, so that 
changes in acceleration are considered to be noise. In a previous study, we
found that this model performed reasonably well on simulated and observed 
pedestrian data.
}
\details{
The constant velocity model is based on the assumption that velocity remains
constant, so that acceleration can be put to 0. This means that we can use 
the following movement equation to model changes in the positions, so that:

\deqn{\begin{\bmatrix}
  x \\
  y \\
\end{bmatrix}_i = \begin{\bmatrix}
  x \\
  y \\
\end{bmatrix}_{i - 1} + \begin{bmatrix}
  v_x \\
  v_y \\
\end{bmatrix} \Delta t}

where \eqn{x} and \eqn{y} represent the x- and y-coordinates on which the 
measurements at time \eqn{t_i} and \eqn{t_{i - 1}} were taken. The variables
\eqn{v_x} and \eqn{v_y} represent the speeds in the x- and y-direction at 
which the system is moving, and \eqn{\Delta t = t_i - t_{i - 1}} represents 
the time between the two measurements. This is the basic equation from which 
the parameters of the Kalman filter are derived.

For the constant velocity model, we keep track of the positional data and 
the respective velocities at a particular time point \eqn{t_i}. This means 
that the latent state \eqn{\mathbf{x}} consists of 4 values on each iteration, 
namely:

\deqn{\mathbf{x}_i = \begin{bmatrix}
  x \\
  y \\
  v_x \\
  v_y \\
\end{bmatrix}_i}

Under this specification, we define the transition matrix \eqn{F} and the 
covariance matrix \eqn{W} of the movement equation as follows:

\deqn{F_i = \begin{bmatrix} 
  1 & 0 & \Delta t & 0 \\  
  0 & 1 & 0 & \Delta t \\
  0 & 0 & 1 & 0 \\
  0 & 0 & 0 & 1 \\
\end{bmatrix}_i}

\deqn{W_i = \begin{bmatrix}
  \Delta t^2 \sigma_{v_x}^2 & 0 & \Delta t \sigma_{v_x}^2 & 0 \\
  0 & \Delta t^2 \sigma_{v_y}^2 & 0 & \Delta t \sigma_{v_y}^2 \\
  \Delta t \sigma_{v_x}^2 & 0 & \sigma_{v_x}^2 & 0 \\
  0 & \Delta t \sigma_{v_y}^2 & 0 & \sigma_{v_y}^2 \\
\end{bmatrix}_i}
In these equations, \eqn{\Delta t} represents the time that has elapsed 
between two observations, so that \eqn{\Delta t = t_i - t_{i - 1}}. The 
variances \eqn{\sigma_{v_x}^2} and \eqn{\sigma_{v_y}^2} capture the 
variation in the speeds \eqn{v_x} and \eqn{v_y}. Within this function, we 
estimate these variances through computing the observed variance in the 
speeds and subtracting the assumed measurement error from it, so that in the 
x-direction, you obtain:

\deqn{\sigma_{v_x}^2 = VAR[v_x]^\text{obs} - 
\frac{2}{E[\Delta t]^2} \sigma_{\epsilon, x}^2}

where \eqn{\sigma_{\epsilon, x}^2} and \eqn{\sigma_{\epsilon, y}^2} are
provided through the \code{error} argument. Note that we assume no 
covariance between the two dimensions in the movement covariance \eqn{W}.

Typically, the movement equation also includes external forces that may 
influence the observed behavior. In the constant velocity model described by
this function, we assume that these parameters have no influence on the 
observed behavior, meaning that we set its parameters \eqn{B} and 
\eqn{\mathbf{u}} to 0.

For the measurement equation, we define the measurement matrix \eqn{H} and 
the measurement error covariance matrix \eqn{R} as:

\deqn{H = \begin{bmatrix}
  1 & 0 & 0 & 0 \\
  0 & 1 & 0 & 0 \\
\end{bmatrix}}

\deqn{R = \begin{bmatrix}
  \sigma_\epsilon^2 & 0 \\
  0 & \sigma_\epsilon^2
\end{bmatrix}}
where \eqn{\sigma_\epsilon^2} is provided through the \code{error}
argument. Importantly, the matrix \eqn{R} is transformed to its Cholesky 
decomposition, as the function \code{\link[denoiser]{kalman_filter}} assumes
the Cholesky decomposition is provided for stability purposes.

Note that these two matrices are time-independent: They are assumed to be 
constant at each iteration (long-term changes over time) and to not depend on 
the time between obsevations \eqn{\Delta t}. Furthermore note that the 
measurement error covariance \eqn{R} can only be defined for those variables
that we have measurements on, namely the x- and y-coordinates. No such error
exists for the speeds \eqn{v_x} and \eqn{v_y}. Similarly, note that the 
measurement matrix \eqn{H} maps the predictions on the latent level to 
predictions the measurement level, where it acknowledges that we only 
measured x- and y-coordinates. 

Finally, we need to define the initial conditions from which the Kalman filter
starts. These initial conditions are taken as the observed initial locations
and speeds (in \eqn{\mathbf{x}_0}) and the observed variances of all 
these variables (in \eqn{P_0}). Note that we provide a covariance matrix 
\eqn{P_0} that is diagonal for simplicity.
}
\examples{
# Generate data for illustration purposes. Movement in circular motion at a
# pace of 1.27m/s with some added noise of SD = 10cm.
# some added noise
angles <- seq(0, 4 * pi, length.out = 100)
coordinates <- 10 * cbind(cos(angles), sin(angles))
coordinates <- coordinates + rnorm(200, mean = 0, sd = 0.1)

data <- data.frame(
  x = coordinates[, 1],
  y = coordinates[, 2],
  time = 1:100
)

# Generate the parameters of the Kalman filter according to the constant 
# velocity model with an assumed measurement error variance of 0.01
constant_velocity(
  data, 
  error = 0.1^2
)

}
