\name{denoiseR-package}
\alias{denoiseR-package}
\alias{denoiseR}
\docType{package}
\title{
Regularized low-rank matrix estimation
}
\description{
The methods implemented allow to recover a low-rank structure from noisy data. In addition, they may be used to estimate the underlying rank.
}
\details{
\tabular{ll}{
Package: \tab denoiseR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-02-09\cr
License: \tab  GPL (>=2)\cr
}
}
\author{
Matan Gavish, Julie Josse, Sylvain Sardy, Stefan Wager \cr
Maintainer: Julie Josse <josse@agrocampus-ouest.fr>\cr
}
\references{
}
\keyword{ package }
\seealso{
URL: http://www.agrocampus-ouest.fr/math/josse\cr
http://web.stanford.edu/~gavish/index.html\cr
http://web.stanford.edu/~swager/research.html\cr
http://www.unige.ch/math/folks/sardy\cr
}
\examples{}
