% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoiser-package.R
\docType{package}
\name{denoiser-package}
\alias{denoiser-package}
\title{denoiser: Noising and denoising spatial data}
\description{
Noiser and denoiser functions to be used for preprocessing spatial data. 
Based on the calibration project of the M4MA team.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ndpvh/denoiser}
  \item Report bugs at \url{https://github.com/ndpvh/denoiser/issues}
}

}
\keyword{internal}
