% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize.R
\name{finalize}
\alias{finalize}
\title{Finalize data after analysis}
\usage{
finalize(data, cols = NULL, .by = NULL)
}
\arguments{
\item{data}{Dataframe that contains information on location (x- and 
y-coordinates) and the time at which the measurement was taken. It is assumed
that this information is contained within the columns \code{"x"}, \code{"y"},
and \code{"time"} respectively.}

\item{cols}{Named vector or named list containing the mapping of the original
column names to the internal ones used within the package. Defaults to 
\code{NULL}, therefore assuming the structure explained in \code{data}.}

\item{.by}{String denoting whether the moving window should be taken with 
respect to a given grouping variable. Defaults to \code{NULL}.}
}
\value{
Adjusted \code{data.frame}
}
\description{
Function that is used internally to change the data back to their original 
format after conducting an analysis, it being either binning 
(\code{\link[denoiser]{bin()}}) or using the Kalman filter 
(\code{\link[denoiser]{kalman_filter()}}). The finalization consists of 
renaming the columns to the user-defined column names instead of the package-
required ones and deleting the grouping variable if it was not originally 
present in the data.
}
\examples{
# Generate data for illustration purposes
#
# Note that for this to work, I need to define the package-required column 
# names
data <- data.frame(
  x = rnorm(100),
  y = rnorm(100),
  time = rep(1:50, times = 2) / 10,
  id = rep(1:2, each = 50)
)

# Prepare the data for analysis
finalize(
  data,
  cols = c(
    "time" = "seconds",
    "x" = "X",
    "y" = "Y"
  ),
  .by = "tag"
)

}
\seealso{
\code{\link[denoiser]{prepare()}}
}
