% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_filter.R
\name{kalman_filter}
\alias{kalman_filter}
\title{Smooth using a Kalman filter}
\usage{
kalman_filter(
  data,
  model = "constant_velocity",
  cols = NULL,
  .by = NULL,
  N_min = 5,
  ...
)
}
\arguments{
\item{data}{Dataframe that contains information on location (x- and 
y-coordinates) and the time at which the measurement was taken. By default, 
[kalman_filter()] will assume that this information is contained within the 
columns `"x"`, `"y"`, and `"time"` respectively. If this isn't the case, 
either change the column names in the data or specify the `cols` argument.}

\item{model}{String denoting which model to use. Defaults to 
`"constant_velocity"`, and is currently the only one that is implemented.}

\item{cols}{Named vector or named list containing the relevant column names
in 'data' if they didn't contain the prespecified column names `"time"`, 
`"x"`, and `"y"`. The labels should conform to these prespecified 
column names and the values given to these locations should contain the 
corresponding column names in that dataset. Defaults to `NULL`, therefore 
assuming the structure explained in `data`.}

\item{.by}{String denoting whether the moving window should be taken with 
respect to a given grouping variable. Defaults to `NULL`.}

\item{N_min}{Integer denoting the minimum number of datapoints that are 
needed to use the Kalman filter. Defaults to \code{5}.}

\item{...}{Additional arguments provided to the loaded model. For more 
information, see [constant_velocity].}
}
\value{
Smoothed dataframe with a similar structure as `data`
}
\description{
This is a higher-level function that will first look at whether the data needs
to be processed by a given variable. Then it will run the Kalman filter on 
these grouped data.
}
\examples{
# Generate data for illustration purposes. Movement in circular motion at a
# pace of 1.27m/s with some added noise of SD = 10cm.
angles <- seq(0, 4 * pi, length.out = 100)
coordinates <- 10 * cbind(cos(angles), sin(angles))
coordinates <- coordinates + rnorm(200, mean = 0, sd = 0.1)

data <- data.frame(
  X = coordinates[, 1],
  Y = coordinates[, 2],
  seconds = rep(1:50, times = 2),
  tag = rep(1:2, each = 50)
)

# Use the Kalman filter with the constant velocity model on these data to 
# filter out the measurement error. Provide an assumed variance of 0.01 
# to this model
kalman_filter(
  data,
  model = "constant_velocity",
  cols = c(
    "time" = "seconds",
    "x" = "X",
    "y" = "Y"
  ),
  .by = "tag",
  error = 0.01
)

}
\seealso{
\code{\link[denoiser]{kf_predict}},
\code{\link[denoiser]{kf_innovation}},
\code{\link[denoiser]{kf_update}}
}
