% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_filter.R
\name{kf_innovation}
\alias{kf_innovation}
\title{Innovation step in the Kalman filter}
\usage{
kf_innovation(z, x, P, H, R)
}
\arguments{
\item{z}{Vector of measured values at time t + 1}

\item{x}{Vector of predicted values of the movement equation at time t + 1.}

\item{P}{Cholesky decomposition of the estimation covariance matrix at time 
t + 1.}

\item{H}{Matrix relating the values in x to the values in y.}

\item{R}{Cholesky decomposition of the measurement noise covariance matrix.#'}
}
\value{
Named list of the innovation (\code{"y"}), the Cholesky decomposition 
of its covariance matrix (\code{"S"}), and the Kalman gain (\code{"K"})
}
\description{
Innovation step in the Kalman filter
}
\details{
In the innovation step, we compute how much error we expect in the measurement
if the movement equation is correct in its prediction. The expected 
"innovation" is defined as:

\deqn{\mathbf{y}_i = \mathbf{z}_i - H \mathbf{x}_i}
where \eqn{\mathbf{z}} is the measurement, \eqn{H} is the measurement matrix, 
and \eqn{\mathbf{x}} is the prediction obtained through the prediction step. 
We also compute the covariance matrix for this innovation, which is defined as:

\deqn{\Sigma_i = H P_i H^T + R}
where \eqn{R} contains the (assumed) measurement covariances. It is 
interesting to note that the value of \eqn{\Sigma} is always symmetric for 
each value of \eqn{H}. Whether its Cholesky decomposition exists, however, 
will depend on both \eqn{H} and \eqn{P}.

In the final step, we compute the Kalman gain using both the covariance of 
the prediction and the total covariance of the measurement process through 
the following equation: 

\deqn{K_i = P_i H^T \Sigma_i^{-1}}
As one may notice, the Kalman gain is a measure of how much we can trust the
prediction, specifically by putting the prediction covariance over the total
variance. It is thus related to a measure of reliability.
}
\seealso{
\code{\link[denoiser]{kalman_filter}},
\code{\link[denoiser]{kf_predict}},
\code{\link[denoiser]{kf_update}}
}
