% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_filter.R
\name{kf_predict}
\alias{kf_predict}
\title{Prediction step in the Kalman filter}
\usage{
kf_predict(
  x0,
  P0,
  F,
  W,
  u = matrix(0, nrow = length(x0), ncol = 1),
  B = matrix(0, nrow = length(u), ncol = length(u))
)
}
\arguments{
\item{x0}{Vector of values of the movement equation at time t.}

\item{P0}{Cholesky decomposition of the covariance of x0 at time t.}

\item{F}{Transition matrix, relating values of X at time t to those at t.}

\item{W}{Cholesky decomposition of the process noise covariance matrix.}

\item{u}{Vector of values for the external variables at time t. By default an 
empty vector.}

\item{B}{Matrix that connects the external variables in u to the values in x.
By default an empty matrix.}
}
\value{
Named list of predicted values for \eqn{\mathbf{x}} (\code{"x}) and 
\eqn{P} (\code{"P"}) at the next time point.
}
\description{
Prediction step in the Kalman filter
}
\details{
In this step, we use the movement equation to predict the new values of x 
based on the initial condition defined by the location \eqn{\mathbf{x}} and 
the covariance \eqn{P} at time \eqn{t_{i - 1}}, using the equation:

\deqn{\mathbf{x}_t = F_i \mathbf{x}_{i - 1} + B_i \mathbf{u}_i + \mathbf{\epsilon}_i}
where \eqn{F} is the movement transition matrix, \eqn{B} scales the external
influences \eqn{\mathbf{u}}, and \eqn{\mathbf{\epsilon}} defines the error, 
and where all terms are defined at the time point \eqn{t_i}. 

The covariance \eqn{P} is also updated based on this equation, representing 
the certainty around the prediction \eqn{\mathbf{x}_i}. This covariance is 
updated through the following equation:

\deqn{P_i = F_i P_{i - 1} F_i^T + W_i}
where \eqn{W} represents the covariance matrix of \eqn{\epsilon}. 

For stability purposes, we use Cholesky decompositions of the covariance 
matrices rather than the actual covariance matrices. This needs to be taken 
into account when making your own model.
}
\seealso{
\code{\link[denoiser]{kalman_filter}},
\code{\link[denoiser]{kf_innovation}},
\code{\link[denoiser]{kf_update}}
}
