% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_filter.R
\name{kf_update}
\alias{kf_update}
\title{Updating step in the Kalman filter}
\usage{
kf_update(x, P, y, H, K)
}
\arguments{
\item{x}{Vector of predicted values of the movement equation at time t + 1.}

\item{P}{Cholesky decomposition of the estimation covariance matrix at time 
t + 1.}

\item{y}{Vector of innovations at time t + 1}

\item{H}{Measurement matrix connecting measurement to movement.}

\item{K}{Kalman gain}
}
\value{
List containing the smoothed value of \eqn{x} (\code{"x"}) together 
with the Cholesky decomposition of its covariance (\code{"P"})
}
\description{
Updating step in the Kalman filter
}
\details{
In the update step, we use the measurement innovation \eqn{\mathbf{y}_i} and 
the predicted value \eqn{\mathbf{x}_i} and combine both in a guess of what 
the latent state of the system should be. For this, we use the following 
equation:

\deqn{\hat{\mathbf{x}}_i = \mathbf{x}_i + K \mathbf{y}}
where \eqn{\mathbf{x}} is the prediction coming from the prediction step, 
\eqn{\mathbf{y}} is the innovation derived in the innovation step, \eqn{K} is
the Kalman gain, and \eqn{\hat{\mathbf{x}}} is the latent state of the system.

We also compute the covariance of this update:

\deqn{\hat{P}_i = (I - K H) P_i}
where I is the identity matrix, \eqn{H} is the measurement matrix, \eqn{K} is 
the Kalman gain, \eqn{P_i} is the covariance of the prediction derived in the 
prediction step, and \eqn{\hat{P}_i} is the estimated certainty around the 
guess \eqn{\hat{\mathbf{x}}_i}. 

Note that the values \eqn{\hat{\mathbf{x}}_i} and \eqn{\hat{P}_i} are used
as initial conditions for the next time step \eqn{t_{i + 1}}, therefore 
serving as input in the prediction step and starting the cycle all over again.
}
\seealso{
\code{\link[denoiser]{kalman_filter}},
\code{\link[denoiser]{kf_predict}},
\code{\link[denoiser]{kf_innovation}}
}
