% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement_models.R
\name{temporal}
\alias{temporal}
\title{Add temporal error to data}
\usage{
temporal(
  data,
  intercept = c(0, 0),
  transition = matrix(c(0.925, 0.085, 0.085, 0.87), nrow = 2, ncol = 2),
  covariance = matrix(c(0.031^2, 0.02, 0.02, 0.027^2), nrow = 2, ncol = 2),
  sampling_rate = 6.13
)
}
\arguments{
\item{data}{A data.frame containing the data on which to base the parameters
of the constant velocity model. This function assumes that this data.frame
contains the columns \code{"time"}, \code{"x"}, and \code{"y"} containing 
the time at which the observed position (x, y) was measured respectively.}

\item{intercept}{Numeric vector with 1 or 2 values containing the intercept 
of the vector autoregressive model.If only 1 number is provided, then this 
value will be assumed to represent the intercept for both the x- and y-
dimensions. Defaults to \code{c(0, 0)}, representing no bias in the system.}

\item{transition}{Numeric matrix of dimensionality \eqn{2 \times 2} having 
the autoregressive effects on the diagonal and crossregressive effects on the
off-diagonal. Defaults to the autoregressive effects being equal to 
\code{0.925} and \code{0.87} for the x- and y-dimension respectively, and 
the cross-regressive effects being equal to \code{0.085} in a symmetric way.
These values represent the means of these parameters as found across the 
four days of the calibration study that informed this project.}

\item{covariance}{Numeric matrix of dimensionality \eqn{2 \times 2}, 
having the variances of the errors for x and y on the diagonal and the 
covariance of the errors on the off-diagonal. Defaults to having the variances
of the x- and y-dimension being equal to \code{0.031^2} and \code{0.027^2}
respectively, and the covariance being equal to \code{0.02}. These values 
represent the means of the respective variances and covariances found 
across the four days of the calibration study that informed this project.}

\item{sampling_rate}{Numeric denoting the sampling rate in Hz. Is used to 
scale the transition matrix according to the actual sampling rate. Defaults to 
\code{6.13}, which is the mean sampling rate for the data obtained across the
four days of the calibration study that informed this project. If you change
the \code{transition} argument, it is recommended to ignore this argument.}
}
\value{
A \code{data.frame} in which the \code{"x"} and \code{"y"} columns 
have additional noise in them.
}
\description{
Error is generated through a vector autoregressive model with its defining 
parameters parameters \code{intercept}, \code{transition}, and 
\code{covariance}, which should be specified by the user. Note that the 
temporal component of this model is defined on the residual level (i.e., 
residuals carry over over time), not on the data level (positions are not 
carried over over time). See the details for more information.
}
\details{
According to this function, the observed positions \eqn{\mathbf{y}_i} at time
\eqn{t_i} contain error that in itself is temporally related, that is carried
over over time. Specifically, we assume that the error 
\eqn{\mathbf{\epsilon}_i} at time \eqn{t_i} depends on the error 
\eqn{\mathbf{\epsilon}_{i - 1}} at previous time \eqn{t_{i - 1}} and a white
noise residual at the same timepoint, as specified in the vector 
autoregressive model. In symbols:

\deqn{\mathbf{\epsilon}_i = \mathbf{\delta} + \Theta \mathbf{\epsilon}_{i - 1} + \mathbf{\omega}_i}

with

\deqn{\mathbf{\omega}_i \sim N(\mathbf{0}, \Sigma)}
where \eqn{\mathbf{\delta}} represents the intercept of the model, 
\eqn{\Theta} is the transition matrix defining the temporal carry-over, and 
\eqn{\Sigma} is the covariance matrix of the white noise residuals 
\eqn{\mathbf{\omega}}. Note that for the vector autoregressive model, the 
mean of the process is defined as:

\deqn{\mathbf{\mu} = (I - \Theta)^{-1} \mathbf{\delta}}
which is useful to keep in mind when specifying the intercept to add bias to 
the system. Furthermore note that the covariance \eqn{\Sigma_\epsilon} of the 
process \eqn{\mathbf{\epsilon}} is defined as:

\deqn{\Sigma_\epsilon = (I - \Theta)^{-2} \Sigma}
which is useful to keep in mind when specifying the covariance matrix 
\eqn{\Sigma}, which is defined on the level of \eqn{\mathbf{\omega}}.

Once the errors \eqn{\mathbf{\epsilon}_i} are defined, we compute the
observed positions \eqn{\mathbf{y}_i} as consisting of a systematic component 
(i.e., the latent positions, provided in \code{data}) and the unsystematic 
error component so that:

\deqn{\mathbf{y}_i = \mathbf{x}_i + \mathbf{\epsilon}_i}
}
\examples{
# Generate data for illustration purposes. Movement in circular motion at a
# pace of 1.27m/s with some added noise of SD = 10cm.
angles <- seq(0, 4 * pi, length.out = 100)
coordinates <- 10 * cbind(cos(angles), sin(angles))

data <- data.frame(
  x = coordinates[, 1],
  y = coordinates[, 2],
  time = 1:100
)

# Add independent noise to these data with an error variance of 0.01 in each
# dimension and a transition matrix containing autoregressive effects of 0.15
temporal(
  data,
  intercept = c(0, 0),
  transition = diag(2) * 0.15,
  covariance = diag(2) * 0.01
) |>
  head()

}
\seealso{
\code{\link[denoiser]{noiser()}}
\code{\link[denoiser]{independent()}}
}
